Develop

+++ Backend/Movieverse.Contracts/DataTransferObjects/Media/MediaDemoDto.cs
@@ -4,7 +4,7 @@ public sealed class MediaDemoDto
 {
 	public Guid Id { get; set; }
 	public string Title { get; set; } = null!;
-	public short Rating { get; set; }
+	public double Rating { get; set; }
 	public Guid? PosterId { get; set; }
 	public Guid? TrailerId { get; set; }
 }
